replace size_t

+++ include/leeter/max_profit_122.hpp
@@ -5,7 +5,7 @@ class Solution {
 public:
   int maxProfit122(vector<int>& prices) {
     int maxProfit = 0;
-    for (size_t i = 1; i < prices.size(); ++i) {
+    for (int i = 1; i < prices.size(); ++i) {
       if (prices[i] > prices[i - 1]) {
         maxProfit += prices[i] - prices[i - 1];
       }
